Start work on adding exporter for vispy volume viewer.
