+++ glue_plotly/html_exporters/qt/__init__.py
@@ -5,3 +5,4 @@ from . import histogram  # noqa
 from . import profile  # noqa
 from . import table  # noqa
 from . import dendrogram  # noqa
+from . import volume  # noqa
