+++ glue_plotly/html_exporters/qt/scatter3d.py
@@ -14,7 +14,8 @@ from glue_qt.viewers.common.tool import Tool
 
 from glue_plotly import PLOTLY_ERROR_MESSAGE, PLOTLY_LOGO
 from glue_plotly.common import data_count, layers_to_export
-from glue_plotly.common.scatter3d import layout_config, traces_for_layer
+from glue_plotly.common.base_3d import layout_config
+from glue_plotly.common.scatter3d import traces_for_layer
 from ... import save_hover, export_dialog
 
 from plotly.offline import plot
